Encapsulates handler function - abstracts into own export
